+++ Afvalwijzer/__init__.py
@@ -3,6 +3,6 @@ __author_email__ = 'bart@bamweb.nl'
 __license__ = 'MIT'
 __maintainer_email__ = 'bart@bamweb.nl'
 __url__ = 'https://github.com/bambam82/afvalwijzer',
-__version__ = '0.2.6'
+__version__ = '0.2.7'
 
 from .Afvalwijzer import Afvalwijzer